Preferences logging & UI panel update-check tweak
Added logging to preferences (in an attempt to figure out why I was always seeing "Failed to check for updates" in my preferences, only to realize I somehow had an outdated version of the file) and changed ui.py to check helpers directly for the current and latest versions rather than using a local variable. Idk if that's even better. Probably?