Stocking Skin Fixes and Better Version Matching
Stocking skin paths will now first try to use the paths stored in Meddle, since those exist now. If none are found (such as if using  TexTools setup, or an older Meddle export) it will fall back to a system of checking for in-use skin textures in a more reliable way than before. This should mean less bibo-skin (and similar) ending up on the legs where they're not meant to be since stockings always use the 'a' type skin.

When looking for variants of the same object / material that should share dyes, shader names in the material names will now be ignored, meaning if an object has multiple materials (such as characterstocking.shpk and characterlegacy.shpk) they will still be able to link together.